Modify master key block mode to CBC over CTR
Modifies the master key block mode from CTR to CBC for greater compatibility with newer JDK (JDK17+) changes that enforce IVs on CTR block mode. Instead of adding a static IV I opted to just change the mode. NOTE: This only affects the encryption and decryption of the keyring. See the README for more details on this change.